requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pytest==8.3.5
factory-boy==3.2.0
faker==8.12.1
pytest-mock==3.10.0
